move jwt specific types to token_verifier_jwt file
